FIX: fix perm issues when installing from build.sh
Fixes permission issues described here: https://github.com/foxcpp/maddy/issues/569

The current configuration allows members of the maddy group to manage maddy
related configuration and read maddy files (the systemd service needs to
be running for the /run/maddy directory to exist).

One concern is that this is achieved using PermissionsStartOnly and
ExecStartPost which appears to be depreciated.